Bias class references have been moved to drp.yaml
