+++ megaradrp/recipes/calibration/__init__.py
@@ -26,5 +26,3 @@ from .flat import FiberFlatRecipe
 from .base import LCB_IFU_StdStarRecipe, FiberMOS_StdStarRecipe
 from .base import SensitivityFromStdStarRecipe, S_And_E_FromStdStarsRecipe
 from .base import BadPixelsMaskRecipe, LinearityTestRecipe
-
-from megaradrp.recipes.calibration.bias import BiasRecipe
